LPD-69154 Rename classes based on BaseCMSTableFDSView class
