+++ modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/frontend/data/set/view/table/ContentAndFileSectionsTableFDSView.java
@@ -27,7 +27,7 @@ import org.osgi.service.component.annotations.Component;
 	service = FDSView.class
 )
 public class ContentAndFileSectionsTableFDSView
-	extends BaseContentsSectionTableFDSView {
+	extends BaseContentsSectionCMSTableFDSView {
 
 	@Override
 	public boolean isDefault() {
